move pkg/config/calculations/calculator.go -> pkg/calculations/calculator.go
